app/backend/requirements.txt
flask==3.0.2
flask-login==0.6.3
google-auth==2.27.0
google-auth-oauthlib==1.2.0
docker==7.1.0
python-dotenv==1.0.1
gunicorn==23.0.0 
pymongo==4.6.2 
pytest==8.0.0
pytest-flask==1.3.0
pytest-mock==3.12.0
mongomock==4.1.2
setuptools==75.8.0
